rename sandbox file
